requirements.txt
asgiref==3.8.1
astroid==2.4.2
Django==4.2.11
django-filter==24.2
djangorestframework==3.11.0
Faker==24.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
python-dateutil==2.9.0.post0
pytz==2020.1
six==1.15.0
sqlparse==0.5.0
toml==0.10.1
typing_extensions==4.11.0
validate-docbr==1.10.0
wrapt==1.12.1
